Update kokkobook.R
hey Tom, I generalized your function distr_shift, so that it can take any value (instead of only values between 13 and 17) and I also added an arrow to the plot (indicating the change in frequencies). I also wrote another function called pop_shift, which is the equivalent of distr_shift, but deals with the numbers of individuals, so you can see population growth at the same time! :)