Update kokkobook.R
I did a very small change to the plotting of the function barnacle, so that when plot=F the line for the next scenario is added to the first plot. So the use would be to run it first with plot=T (default), and then add scenarios with plot=F, allowing us to arrive at a figure like the one in the book.